Implement Web API service and Queue/Storage mechanism
The Web application should work with different SDK client's with help of REST API.
Web application accept the HTTP request and save the payload to queue for further processing/aggregation to achieve the main goal (accept thousands incoming requests).
To achieve the main goal this commit adds a few packages to work with Queue and DataBase engines.
The Web application should work with different SDK client's with help of REST API this commit adds support of REST routes and basic validation for input data.